Ensure `key` isn't overwritten, when we cleanup the device_cache.
When `device_cache` hits `_max_allocations`, we evict the overflowing items and rebuild the table and cache.

However, we were overwriting the `key` variable, which was causing weirdness down the road.

Eventually, our cache ended up in a state, where there was only a single key in `_cache`, that corresponded to ALL indices.

In the specific test case, the last key was 4x132. This bug caused device_arrays that had shape != 4x132 to be returned.

It also resulted in tons of cache misses, that should have been cache hits. This may have performance implications for our benchmarks.