Handle prefill EOT and remove tokenizer padding
Needed to remove padding for batch encode. Encoding does not include the
sequence length for batch encoding so this would result in end paddings
of the sequence.

Also - if the first token encountered was a EOS token then stepping
would fail. Reworked initialization / errors to better handle this and
have a uniform stepping location of page ids.