Output QType needs to be set as `torch.float16` for KV Cache (#2234)
The qscale type is being used to infer the q_output dtype which results
in `torch.float32` while `torch.float16` is correct. Hardcoded for now
but eventually `attn_dtype` should be plumbed.

This occurs when exporting with the torch operator and using an `fp8`
cache but dequant to `fp16` for attention. Guarantee correct attention
dtypes avoids running the `torch` operator with mixed types.