[Shortfin LLM] Fix extra beams (#2282)
Stacked on #2276 

This fixes an issue that we were seeing where sometimes, you would get a
response w/ for example 3 beams, when you requested 2. Eventually we can
modify this and add another `num_responses` arg, which is what other
servers tend to do.

In the meantime, this seems like the intended result.

---------

Co-authored-by: lisaliu1 <lisal@amd.com>