[sharktank] Fix dispatch resolution for ops.linear when all tensors are replicated (#2302)
Fixes #2300.

Delegate to the standard handling of the trivially replicable mechanism
when all args are replicated.
The custom sharded implementation should be bypassed in the
all-replicated case.

Also do not mix sharded and unsharded tensors in the op. We used to
allow this before, but now we want all tensors to be either sharded.