[sharktank] Implement unpack for replicated quantized tensors (#2310)
Fixes https://github.com/nod-ai/shark-ai/issues/2299.

The unpack for quantized replicated tensors should not be trivially
replicable. It returns a layout that is not a tensor. But rather
contains tensors(planes).

A sharded quantized tensor is a collection of tensors that each is
backed by unsharded layout.
But then we must construct a layout that is composed by sharded planes.
This transformation is not trivially replicable.

Potentially we could have in the future a more general approach for
automatic op replication where we describe how to combine the results of
an all-replicated tensor args op. E.g. how to combine a bunch of
quantized layouts.
Maybe we should be able to express replicated things (e.g. layouts) and
not just tensors.