[Fusilli] Use LLVM's FileCheck (#2364)
The python [filecheck ](https://github.com/AntonLydike/filecheck) does
have sharp edges in terms of supported options / maturity. Hit an issue
today with `--dump-input` not working as expected and exposing an odd
arg parsing bug. Based on
https://github.com/AntonLydike/filecheck/pull/36 it looks like not all
options are supported.

LLVM's FileCheck is pre-packaged in the clang distribution we depend on
so it wasn't too difficult to switch over.