[Fusilli] Fix to allow benchmark tests to be run concurrently (`-j N`) (#2470)
Fixes https://github.com/nod-ai/shark-ai/issues/2442 by parameterizing
the graphName for the benchmark driver with the various conv variant
options provided through CLI. This should address the immediate issue
and allow running `ctest -j <N>`...

Since graphName is expected to be unique, and is used for caching, the
user code can just do any uniquefying on their end (e.g. with a random
nonce or any logic of their choosing) - we don't have to do it in the
library itself.