[LLM Server] Fix deadlock in non-resizable fiber pool (#2517)
- Through testing adding an `invocation_fiber_pool` to enable
overlapping executions, found a bug in `non-resizable FiberPools`.
- Since our `await self.__index_queue.get()` in `async def get` is
inside of the `lock's context manager`, we would never actually release
the lock to allow the currently allocated fibers to acquire it. This
means they can never return their fiber to the pool, which means the
`self.__index_queue.get()` coroutine will never complete, which then
means that we can never release the lock, causing a deadlock.
- Added a unit test for this case