[Fusilli] Create FUSILLI_REQUIRE_OK macro (#2556)
This odd [CI
failure](https://github.com/nod-ai/shark-ai/actions/runs/18548883676/job/52872578004)
might indicate a race on the compile cache (though I don't see how) or a
sporadic device issue. It's hard to tell from the logs, and I haven't
been able to reproduce this locally. This PR adds some logging so
hopefully if this happens in CI again we can get a better picture of
what's going on.